Rewrite in Haskell
--- Motivation

Q: Why a rewrite?
A: I want jsonwatch to be simpler to deploy -- ideally a single
static binary. It currently has under 300 lines of Python code,
which makes it possible to rewrite in a reasonable time in a
language where building binaries is a first-class workflow.

Q: Why rewrite in Haskell and not _?
A: To learn Haskell. :-)

--- Functionality changes

The rewritten version takes the same command line options and
generally works the same. However, there are some changes, which
are detailed below.

* Format dates differently: no milliseconds, include the time zone.

* Do not fail on invalid JSON data. Instead, treat invalid data as
an empty document for the purposes of diffing.

* Reorder diff lines: show the changed values first, then the
removed keys, then the added keys.

v0.3.0